customer-service/requirements.txt
flask
mysql-connector-python
confluent-kafka
python-dotenv